pyproject.toml
[tool.poetry]
name = "news-manager"
version = "0.1.0"
description = ""
authors = ["Cácio Lucas <cclucas060901@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
dj-database-url = "^2.3.0"
django-cors-headers = "^4.7.0"
pillow = "^11.1.0"
django-summernote = "^0.8.20.0"
django-prose-editor = "^0.10.3"
django-taggit = "^6.1.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
